change projections in emissions.py
i decided to do the projections earlier, right after we load the emissions file. i also deleted the additional projection in buffer_emis afterwards. i've also edited the code for the hole since i noticed the warning was being triggered for 140.0 in the file where there was supposed to be no hole, so i'm assuming we want to exclude 140!